Fix build with gcc -m32 on x86-64-linux.
There were three problems.  The first was warnings in
sha512_finish_ctx caused by uint64_t being bigger than size_t; there's
no good reason why sha512 ctx->buflen needs to be a uint64_t, shrink
it to uint32_t like it is for sha256.  The other problems had to do
with symbol versioning: the test for x32 in libcrypt.minver got a
false positive on gcc's -m32 mode (which is effectively cross
compilation to i686, but config.guess doesn't know that) and
test-symbols-static didn't know to ignore compiler-emitted symbols
starting with two underscores.

	* alg-sha512.h (struct sha512_ctx): Make buflen field uint32_t.
	* alg-sha512.c (sha512_finish_ctx): Make bytes variable uint32_t.
        (sha512_finish_bytes): Make left_over and add variables uint32_t.

	* test-symbols-static (list_library_globals): Also ignore symbols
	beginning with __ or _[A-Y].  (_Z is the prefix for C++ mangled
	names, so we don't want to ignore that one.)

	* libcrypt.minver: Tighten test for x32.
	* Makefile.am (CONFIG_STATUS_DEPENDENCIES): Add libcrypt.minver.

+++ alg-sha512.c
@@ -201,7 +201,7 @@ void *
 sha512_finish_ctx (struct sha512_ctx *ctx, void *resbuf)
 {
   /* Take yet unprocessed bytes into account.  */
-  uint64_t bytes = ctx->buflen;
+  uint32_t bytes = ctx->buflen;
   size_t pad;
   unsigned int i;
   unsigned char *rp = resbuf;
@@ -242,8 +242,8 @@ sha512_process_bytes (const void *buffer, size_t len, struct sha512_ctx *ctx)
      both inputs first.  */
   if (ctx->buflen != 0)
     {
-      size_t left_over = ctx->buflen;
-      size_t add = 256 - left_over > len ? len : 256 - left_over;
+      uint32_t left_over = ctx->buflen;
+      uint32_t add = 256 - left_over > len ? (uint32_t)len : 256 - left_over;
 
       memcpy (&ctx->buffer[left_over], buffer, add);
       ctx->buflen += add;
@@ -283,6 +283,6 @@ sha512_process_bytes (const void *buffer, size_t len, struct sha512_ctx *ctx)
           left_over -= 128;
           memcpy (ctx->buffer, &ctx->buffer[128], left_over);
         }
-      ctx->buflen = left_over;
+      ctx->buflen = (uint32_t)left_over;
     }
 }
